Expose the topo/geojson properties as extra models and take a stab at parsing the area infos
The properties are part of the geo/topojson spec but that information is only included there
and well structured, so provide an API for parsing/exposing it.

For the area infos: even though I haven't seen them in the wild this tries to parse them.
We'll see if it is wrong when parsing fails.